fix:sdk location provider
