+++ pkg/plugin/metrics/proxy.go
