+++ pkg/plugin/metrics/statreporter.go
