+++ pkg/plugin/register/plugins.go
@@ -25,10 +25,10 @@ import (
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/healthcheck"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/loadbalancer"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/localregistry"
+	_ "github.com/polarismesh/polaris-go/pkg/plugin/metrics"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/ratelimiter"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/serverconnector"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/servicerouter"
-	_ "github.com/polarismesh/polaris-go/pkg/plugin/statreporter"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/subscribe"
 	_ "github.com/polarismesh/polaris-go/pkg/plugin/weightadjuster"
 	_ "github.com/polarismesh/polaris-go/plugin/alarmreporter/file"
@@ -44,6 +44,8 @@ import (
 	_ "github.com/polarismesh/polaris-go/plugin/loadbalancer/weightedrandom"
 	_ "github.com/polarismesh/polaris-go/plugin/localregistry/inmemory"
 	_ "github.com/polarismesh/polaris-go/plugin/logger/zaplog"
+	_ "github.com/polarismesh/polaris-go/plugin/metrics/prometheus"
+	_ "github.com/polarismesh/polaris-go/plugin/metrics/pushgateway"
 	_ "github.com/polarismesh/polaris-go/plugin/ratelimiter/reject"
 	_ "github.com/polarismesh/polaris-go/plugin/ratelimiter/unirate"
 	_ "github.com/polarismesh/polaris-go/plugin/serverconnector/grpc"
@@ -53,8 +55,6 @@ import (
 	_ "github.com/polarismesh/polaris-go/plugin/servicerouter/nearbybase"
 	_ "github.com/polarismesh/polaris-go/plugin/servicerouter/rulebase"
 	_ "github.com/polarismesh/polaris-go/plugin/servicerouter/setdivision"
-	_ "github.com/polarismesh/polaris-go/plugin/statreporter/prometheus"
-	_ "github.com/polarismesh/polaris-go/plugin/statreporter/pushgateway"
 	_ "github.com/polarismesh/polaris-go/plugin/subscribe/localchannel"
 	_ "github.com/polarismesh/polaris-go/plugin/weightadjuster/ratedelay"
 )
