+++ plugin/metrics/prometheus/addons/max_gauge.go
