+++ plugin/metrics/prometheus/addons/max_gauge_test.go
