+++ plugin/metrics/prometheus/prometheus_handler.go
@@ -14,52 +14,73 @@
 // specific language governing permissionsr and limitations under the License.
 //
 
-package pushgateway
+package prometheus
 
 import (
 	"fmt"
+	"net"
 	"net/http"
+	"sync"
 
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 
 	"github.com/polarismesh/polaris-go/pkg/log"
 	"github.com/polarismesh/polaris-go/pkg/model"
 	"github.com/polarismesh/polaris-go/pkg/plugin"
-	"github.com/polarismesh/polaris-go/plugin/statreporter/prometheus/addons"
+	"github.com/polarismesh/polaris-go/plugin/metrics/prometheus/addons"
 )
 
-// PushgatewayHandler handler for prometheus
-type PushgatewayHandler struct {
+// PrometheusHandler handler for prometheus
+type PrometheusHandler struct {
 	// prometheus的metrics注册
 	registry *prometheus.Registry
 	// metrics的 http handler
-	handler         http.Handler
-	cfg             *Config
+	handler http.Handler
+	cfg     *Config
+	//
 	metricVecCaches map[string]prometheus.Collector
+	ln              net.Listener
 	bindIP          string
+	port            int
 }
 
-func newHandler(ctx *plugin.InitContext) (*PushgatewayHandler, error) {
-	p := &PushgatewayHandler{}
+func newHandler(ctx *plugin.InitContext) (*PrometheusHandler, error) {
+	p := &PrometheusHandler{}
 	return p, p.init(ctx)
 }
 
-func (p *PushgatewayHandler) init(ctx *plugin.InitContext) error {
+func (p *PrometheusHandler) init(ctx *plugin.InitContext) error {
 	cfgValue := ctx.Config.GetGlobal().GetStatReporter().GetPluginConfig(PluginName)
 	if cfgValue != nil {
 		p.cfg = cfgValue.(*Config)
 	}
-	p.bindIP = ctx.Config.GetGlobal().GetAPI().GetBindIP()
+
 	p.metricVecCaches = make(map[string]prometheus.Collector)
 	p.registry = prometheus.NewRegistry()
 	if err := p.registerMetrics(); err != nil {
 		return err
 	}
 
+	p.bindIP = p.cfg.IP
+	if p.bindIP == "" {
+		p.bindIP = ctx.Config.GetGlobal().GetAPI().GetBindIP()
+	}
+	p.port = p.cfg.Port
+
+	p.handler = &metricsHttpHandler{
+		promeHttpHandler: promhttp.HandlerFor(p.registry, promhttp.HandlerOpts{}),
+		lock:             &sync.RWMutex{},
+	}
+
+	if ctx.Config.GetGlobal().GetStatReporter().IsEnable() {
+		p.runInnerMetricsWebServer()
+	}
+
 	return nil
 }
 
-func (p *PushgatewayHandler) registerMetrics() error {
+func (p *PrometheusHandler) registerMetrics() error {
 	for _, desc := range metrcisDesces {
 		var collector prometheus.Collector
 		switch desc.MetricType {
@@ -100,7 +121,7 @@ func (p *PushgatewayHandler) registerMetrics() error {
 }
 
 // ReportStat 上报采集指标到 prometheus，这里只针对部分 model.InstanceGauge 的实现做处理
-func (p *PushgatewayHandler) ReportStat(metricsType model.MetricType, metricsVal model.InstanceGauge) error {
+func (p *PrometheusHandler) ReportStat(metricsType model.MetricType, metricsVal model.InstanceGauge) error {
 	switch metricsType {
 	case model.ServiceStat:
 		val, ok := metricsVal.(*model.ServiceCallResult)
@@ -121,7 +142,41 @@ func (p *PushgatewayHandler) ReportStat(metricsType model.MetricType, metricsVal
 	return nil
 }
 
-func (p *PushgatewayHandler) handleServiceGauge(metricsType model.MetricType, val *model.ServiceCallResult) {
+// runInnerMetricsWebServer 启动用于 prometheus 主动拉取的 http-server，如果端口设置为负数，则不启用
+func (p *PrometheusHandler) runInnerMetricsWebServer() {
+	if p.port < 0 {
+		return
+	}
+
+	ln, err := net.Listen("tcp", fmt.Sprintf("%s:%d", p.bindIP, p.cfg.Port))
+	if err != nil {
+		log.GetBaseLogger().Errorf("start metrics http-server fail: %v", err)
+		p.port = -1
+		return
+	}
+
+	p.ln = ln
+	p.port = ln.Addr().(*net.TCPAddr).Port
+
+	go func() {
+		log.GetBaseLogger().Infof("start metrics http-server address : %s", fmt.Sprintf("%s:%d", p.bindIP, p.port))
+		if err := http.Serve(ln, p.GetHttpHandler()); err != nil {
+			log.GetBaseLogger().Errorf("start metrics http-server fail : %s", err)
+			return
+		}
+	}()
+}
+
+func (p *PrometheusHandler) exportSuccess() bool {
+	return p.port > 0
+}
+
+// GetHttpHandler 获取 handler
+func (p *PrometheusHandler) GetHttpHandler() http.Handler {
+	return p.handler
+}
+
+func (p *PrometheusHandler) handleServiceGauge(metricsType model.MetricType, val *model.ServiceCallResult) {
 	labels := p.convertInsGaugeToLabels(val)
 
 	total := p.metricVecCaches[MetricsNameUpstreamRequestTotal].(*prometheus.CounterVec)
@@ -147,7 +202,7 @@ func (p *PushgatewayHandler) handleServiceGauge(metricsType model.MetricType, va
 	}
 }
 
-func (p *PushgatewayHandler) handleRateLimitGauge(metricsType model.MetricType, val *model.RateLimitGauge) {
+func (p *PrometheusHandler) handleRateLimitGauge(metricsType model.MetricType, val *model.RateLimitGauge) {
 	labels := p.convertRateLimitGaugeToLabels(val)
 
 	total := p.metricVecCaches[MetricsNameRateLimitRequestTotal].(*prometheus.CounterVec)
@@ -164,7 +219,7 @@ func (p *PushgatewayHandler) handleRateLimitGauge(metricsType model.MetricType, 
 	}
 }
 
-func (p *PushgatewayHandler) handleCircuitBreakGauge(metricsType model.MetricType, val *model.CircuitBreakGauge) {
+func (p *PrometheusHandler) handleCircuitBreakGauge(metricsType model.MetricType, val *model.CircuitBreakGauge) {
 	labels := p.convertCircuitBreakGaugeToLabels(val)
 
 	open := p.metricVecCaches[MetricsNameCircuitBreakerOpen].(*prometheus.GaugeVec)
@@ -186,7 +241,7 @@ func (p *PushgatewayHandler) handleCircuitBreakGauge(metricsType model.MetricTyp
 	}
 }
 
-func (p *PushgatewayHandler) convertInsGaugeToLabels(val *model.ServiceCallResult) map[string]string {
+func (p *PrometheusHandler) convertInsGaugeToLabels(val *model.ServiceCallResult) map[string]string {
 	labels := make(map[string]string)
 
 	for label, supplier := range InstanceGaugeLabelOrder {
@@ -197,7 +252,7 @@ func (p *PushgatewayHandler) convertInsGaugeToLabels(val *model.ServiceCallResul
 	return labels
 }
 
-func (p *PushgatewayHandler) convertRateLimitGaugeToLabels(val *model.RateLimitGauge) map[string]string {
+func (p *PrometheusHandler) convertRateLimitGaugeToLabels(val *model.RateLimitGauge) map[string]string {
 	labels := make(map[string]string)
 
 	for label, supplier := range RateLimitGaugeLabelOrder {
@@ -206,7 +261,7 @@ func (p *PushgatewayHandler) convertRateLimitGaugeToLabels(val *model.RateLimitG
 	return labels
 }
 
-func (p *PushgatewayHandler) convertCircuitBreakGaugeToLabels(val *model.CircuitBreakGauge) map[string]string {
+func (p *PrometheusHandler) convertCircuitBreakGaugeToLabels(val *model.CircuitBreakGauge) map[string]string {
 	labels := make(map[string]string)
 
 	for label, supplier := range CircuitBreakerGaugeLabelOrder {
@@ -216,6 +271,11 @@ func (p *PushgatewayHandler) convertCircuitBreakGaugeToLabels(val *model.Circuit
 }
 
 // Close the prometheus handler
-func (p *PushgatewayHandler) Close() error {
+func (p *PrometheusHandler) Close() error {
+	if p.ln != nil {
+		if err := p.ln.Close(); err != nil {
+			return err
+		}
+	}
 	return nil
 }
