+++ plugin/metrics/pushgateway/addons/max_gauge.go
