+++ plugin/metrics/pushgateway/addons/max_gauge_test.go
