+++ plugin/metrics/pushgateway/pushgateway_reporter.go
@@ -15,32 +15,35 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package prometheus
+package pushgateway
 
 import (
-	"net/http"
-	"sync"
+	"time"
 
+	"github.com/prometheus/client_golang/prometheus/push"
+
+	"github.com/polarismesh/polaris-go/pkg/log"
 	"github.com/polarismesh/polaris-go/pkg/model"
 	"github.com/polarismesh/polaris-go/pkg/plugin"
 	"github.com/polarismesh/polaris-go/pkg/plugin/common"
-	"github.com/polarismesh/polaris-go/pkg/plugin/statreporter"
+	statreporter "github.com/polarismesh/polaris-go/pkg/plugin/metrics"
 )
 
 const (
 	// PluginName is the name of the plugin.
-	PluginName string = "prometheus"
+	PluginName      = "pushgateway"
+	_defaultJobName = "polaris-client"
 )
 
-var _ statreporter.StatReporter = (*PrometheusReporter)(nil)
+var _ statreporter.StatReporter = (*PushgatewayReporter)(nil)
 
 // init 注册插件.
 func init() {
-	plugin.RegisterPlugin(&PrometheusReporter{})
+	plugin.RegisterPlugin(&PushgatewayReporter{})
 }
 
-// PrometheusReporter is a prometheus reporter.
-type PrometheusReporter struct {
+// PushgatewayReporter is a prometheus reporter.
+type PushgatewayReporter struct {
 	*plugin.PluginBase
 	*common.RunContext
 	// 本插件的配置
@@ -52,21 +55,22 @@ type PrometheusReporter struct {
 	// 插件工厂
 	plugins plugin.Supplier
 	// prometheus的metrics注册
-	handler *PrometheusHandler
+	handler *PushgatewayHandler
+	ticker  *time.Ticker
 }
 
 // Type 插件类型.
-func (s *PrometheusReporter) Type() common.Type {
+func (s *PushgatewayReporter) Type() common.Type {
 	return common.TypeStatReporter
 }
 
 // Name 插件名，一个类型下插件名唯一.
-func (s *PrometheusReporter) Name() string {
+func (s *PushgatewayReporter) Name() string {
 	return PluginName
 }
 
 // Init 初始化插件.
-func (s *PrometheusReporter) Init(ctx *plugin.InitContext) error {
+func (s *PushgatewayReporter) Init(ctx *plugin.InitContext) error {
 	s.RunContext = common.NewRunContext()
 	s.globalCtx = ctx.ValueCtx
 	s.plugins = ctx.Plugins
@@ -76,36 +80,34 @@ func (s *PrometheusReporter) Init(ctx *plugin.InitContext) error {
 		return err
 	}
 	s.handler = handler
+	cfgValue := ctx.Config.GetGlobal().GetStatReporter().GetPluginConfig(PluginName)
+	s.cfg = cfgValue.(*Config)
+	if cfgValue != nil {
+		s.runPushMetrics(cfgValue.(*Config).PushInterval)
+	}
 	return nil
 }
 
 // ReportStat 报告统计数据.
-func (s *PrometheusReporter) ReportStat(metricType model.MetricType, metricsVal model.InstanceGauge) error {
+func (s *PushgatewayReporter) ReportStat(metricType model.MetricType, metricsVal model.InstanceGauge) error {
 	return s.handler.ReportStat(metricType, metricsVal)
 }
 
 // Info 插件信息.
-func (s *PrometheusReporter) Info() model.StatInfo {
-	if !s.handler.exportSuccess() {
-		return model.StatInfo{}
-	}
-	return model.StatInfo{
-		Target:   s.Name(),
-		Port:     uint32(s.handler.port),
-		Path:     "/metrics",
-		Protocol: "http",
-	}
+func (s *PushgatewayReporter) Info() model.StatInfo {
+	return model.StatInfo{}
 }
 
 // Destroy .销毁插件.
-func (s *PrometheusReporter) Destroy() error {
-	err := s.PluginBase.Destroy()
-	if err != nil {
+func (s *PushgatewayReporter) Destroy() error {
+	if err := s.PluginBase.Destroy(); err != nil {
 		return err
 	}
-	err = s.RunContext.Destroy()
-	if err != nil {
+	if err := s.RunContext.Destroy(); err != nil {
 		return err
+	}
+	if s.ticker != nil {
+		s.ticker.Stop()
 	}
 
 	if s.handler != nil {
@@ -117,14 +119,17 @@ func (s *PrometheusReporter) Destroy() error {
 	return nil
 }
 
-type metricsHttpHandler struct {
-	promeHttpHandler http.Handler
-	lock             *sync.RWMutex
-}
-
-// ServeHTTP 提供 prometheus http 服务.
-func (p *metricsHttpHandler) ServeHTTP(writer http.ResponseWriter, request *http.Request) {
-	p.lock.RLock()
-	defer p.lock.RUnlock()
-	p.promeHttpHandler.ServeHTTP(writer, request)
+// runPushMetrics 指标推送
+func (s *PushgatewayReporter) runPushMetrics(interval time.Duration) {
+	s.ticker = time.NewTicker(interval)
+	go func() {
+		for range s.ticker.C {
+			if err := push.
+				New(s.cfg.Address, _defaultJobName).
+				Gatherer(s.handler.registry).
+				Push(); err != nil {
+				log.GetBaseLogger().Errorf("push metrics to pushgateway fail: %s", err.Error())
+			}
+		}
+	}()
 }
